+++ src/NexusMods.DataModel/DataModelSettings.cs
@@ -20,6 +20,11 @@ public record DataModelSettings : ISettings
     /// Path of the file which contains the backing data store or database.
     /// </summary>
     public ConfigurablePath DataStoreFilePath { get; set; }
+    
+    /// <summary>
+    /// Path of the folder containing the MnemonicDB database.
+    /// </summary>
+    public ConfigurablePath MnemonicDBPath { get; set; }
 
     /// <summary>
     /// Preconfigured locations [full paths] where mod archives can/will be stored.
@@ -53,6 +58,7 @@ public record DataModelSettings : ISettings
         return new DataModelSettings
         {
             DataStoreFilePath = new ConfigurablePath(baseKnownPath, $"{baseDirectoryName}/DataModel.sqlite"),
+            MnemonicDBPath = new ConfigurablePath(baseKnownPath, $"{baseDirectoryName}/MnemonicDB.rocksdb"),
             ArchiveLocations = [
                 new ConfigurablePath(baseKnownPath, $"{baseDirectoryName}/Archives"),
             ],