+++ src/NexusMods.DataModel/Services.cs
@@ -1,5 +1,6 @@
 using System.Text.Json.Serialization;
 using Microsoft.Extensions.DependencyInjection;
+using NexusMods.Abstractions.Settings;
 using NexusMods.Abstractions.Diagnostics;
 using NexusMods.Abstractions.DiskState;
 using NexusMods.Abstractions.Games.Loadouts;
@@ -10,7 +11,7 @@ using NexusMods.Abstractions.Loadouts;
 using NexusMods.Abstractions.Messaging;
 using NexusMods.Abstractions.Serialization;
 using NexusMods.Abstractions.Serialization.ExpressionGenerator;
-using NexusMods.Abstractions.Settings;
+using NexusMods.DataModel.Attributes;
 using NexusMods.DataModel.CommandLine.Verbs;
 using NexusMods.DataModel.Diagnostics;
 using NexusMods.DataModel.JsonConverters;
@@ -20,6 +21,11 @@ using NexusMods.DataModel.Settings;
 using NexusMods.DataModel.Sorting;
 using NexusMods.DataModel.TriggerFilter;
 using NexusMods.Extensions.DependencyInjection;
+using NexusMods.MnemonicDB;
+using NexusMods.MnemonicDB.Abstractions;
+using NexusMods.MnemonicDB.Storage;
+using NexusMods.MnemonicDB.Storage.Abstractions;
+using NexusMods.Paths;
 
 namespace NexusMods.DataModel;
 
@@ -32,9 +38,47 @@ public static class Services
     /// </summary>
     public static IServiceCollection AddDataModel(this IServiceCollection coll)
     {
+
+        coll.AddMnemonicDB();
+        coll.AddMnemonicDBStorage();
+
         coll.AddSettings<DataModelSettings>();
         coll.AddSettingsStorageBackend<DataStoreSettingsBackend>(isDefault: true);
 
+        coll.AddSingleton<MnemonicDB.Storage.InMemoryBackend.Backend>();
+        coll.AddSingleton<MnemonicDB.Storage.RocksDbBackend.Backend>();
+
+        coll.AddSingleton<DatomStoreSettings>(sp =>
+            {
+                var fileSystem = sp.GetRequiredService<IFileSystem>();
+                var settingsManager = sp.GetRequiredService<ISettingsManager>();
+                var settings = settingsManager.Get<DataModelSettings>();
+                return new DatomStoreSettings
+                {
+                    Path = settings.MnemonicDBPath.ToPath(fileSystem),
+                };
+            }
+        );
+        
+        coll.AddSingleton<IStoreBackend>(sp =>
+        {
+            var settingsManager = sp.GetRequiredService<ISettingsManager>();
+            var settings = settingsManager.Get<DataModelSettings>();
+            if (settings.UseInMemoryDataModel)
+            {
+                return sp.GetRequiredService<MnemonicDB.Storage.InMemoryBackend.Backend>();
+            }
+            else
+            {
+                var datomStoreSettings = sp.GetRequiredService<DatomStoreSettings>();
+
+                if (!datomStoreSettings.Path.DirectoryExists()) 
+                    datomStoreSettings.Path.CreateDirectory();
+                
+                return sp.GetRequiredService<MnemonicDB.Storage.RocksDbBackend.Backend>();
+            }
+        });
+        
         coll.AddSingleton<MessageBus>();
         coll.AddSingleton(typeof(IMessageConsumer<>), typeof(MessageConsumer<>));
         coll.AddSingleton(typeof(IMessageProducer<>), typeof(MessageProducer<>));
@@ -56,7 +100,11 @@ public static class Services
         coll.AddAllSingleton<IFileHashCache, FileHashCache>();
         coll.AddAllSingleton<IArchiveInstaller, ArchiveInstaller>();
         coll.AddAllSingleton<IToolManager, ToolManager>();
+        
+        // Disk State Registry
         coll.AddAllSingleton<IDiskStateRegistry, DiskStateRegistry>();
+        coll.AddAttributeCollection(typeof(DiskState));
+        
         coll.AddAllSingleton<IApplyService, ApplyService>();
 
         coll.AddSingleton<ITypeFinder>(_ => new AssemblyTypeFinder(typeof(Services).Assembly));
@@ -65,7 +113,7 @@ public static class Services
         // Diagnostics
         coll.AddAllSingleton<IDiagnosticManager, DiagnosticManager>();
         coll.AddSettings<DiagnosticSettings>();
-
+        
         // Verbs
         coll.AddLoadoutManagementVerbs()
             .AddToolVerbs()